# swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TodoListApp API",
    "version": "1.0.0",
    "description": "TodoListApp REST API 명세서입니다. 사용자 인증 기반의 개인 할일 관리 플랫폼으로, 카테고리별 할일 생성·조회·수정·삭제·완료 처리를 지원합니다.\n\n## 인증 방식\n모든 보호된 엔드포인트는 `Authorization: Bearer <accessToken>` 헤더가 필요합니다.\n\n## 에러 응답 형식\n```json\n{ \"error\": { \"code\": \"ERROR_CODE\", \"message\": \"에러 설명\" } }\n```\n\n## 기본 카테고리\n시스템에서 제공하는 기본 카테고리(`업무`, `개인`, `쇼핑`)는 모든 사용자가 공유하며, 삭제할 수 없습니다.",
    "contact": {
      "name": "TodoListApp",
      "email": "hoppang250@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "로컬 개발 서버"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "로그인 시 발급된 Access Token을 입력하세요. 유효기간: 1시간"
      }
    },
    "schemas": {
...
      "Todo": {
        "type": "object",
        "description": "할일 정보. overdue 필드는 dueDate가 오늘 이전이고 isCompleted가 false인 경우 true로 계산됩니다 (DB 저장 안 함).",
        "properties": {
          "todoId": {
            "type": "string",
            "format": "uuid",
            "description": "할일 고유 식별자",
            "example": "550e8400-e29b-41d4-a716-446655440100"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "description": "소유 사용자 ID",
            "example": "550e8400-e29b-41d4-a716-446655440001"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "description": "분류 카테고리 ID",
            "example": "550e8400-e29b-41d4-a716-446655440010"
          },
          "title": {
            "type": "string",
            "maxLength": 255,
            "description": "할일 제목 (필수, 최대 255자)",
            "example": "주간 보고서 작성"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "할일 상세 설명 (선택)",
            "example": "월요일 오전 9시까지 제출"
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "description": "종료 예정일 (선택, YYYY-MM-DD 형식)",
            "example": "2026-05-20"
          },
          "isCompleted": {
            "type": "boolean",
            "description": "완료 여부. true이면 완료, false이면 미완료",
            "example": false
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "실제 완료 일시. 완료 처리 시 기록되며 완료 취소 시 null로 초기화",
            "example": null
          },
          "overdue": {
            "type": "boolean",
            "description": "기한 초과 여부. dueDate < 오늘 날짜 AND isCompleted = false인 경우 true (계산 속성)",
            "example": false
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "할일 생성 일시",
            "example": "2026-05-13T10:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "할일 최종 수정 일시",
            "example": "2026-05-13T10:30:00.000Z"
          }
        },
        "required": [
          "todoId",
          "userId",
          "categoryId",
          "title",
          "description",
          "dueDate",
          "isCompleted",
          "completedAt",
          "overdue",
          "createdAt",
          "updatedAt"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "description": "표준 에러 응답 형식",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "에러 코드",
                "enum": [
                  "UNAUTHORIZED",
                  "INVALID_TOKEN",
                  "INVALID_INPUT",
                  "DUPLICATE_EMAIL",
                  "NOT_FOUND",
                  "FORBIDDEN",
                  "INTERNAL_SERVER_ERROR"
                ],
                "example": "INVALID_INPUT"
              },
              "message": {
                "type": "string",
                "description": "사람이 읽을 수 있는 에러 설명",
                "example": "이메일과 비밀번호는 필수 입력 항목입니다."
              }
            },
            "required": ["code", "message"]
          }
        },
        "required": ["error"]
      },
...
      "CreateTodoRequest": {
        "type": "object",
        "description": "할일 생성 요청 본문. title과 categoryId는 필수입니다.",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "description": "할일 제목 (필수, 최대 255자)",
            "example": "주간 보고서 작성"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "description": "할일을 분류할 카테고리 ID (필수)",
            "example": "550e8400-e29b-41d4-a716-446655440010"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "할일 상세 설명 (선택)",
            "example": "월요일 오전 9시까지 팀장에게 제출"
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "description": "종료 예정일 (선택, YYYY-MM-DD 형식)",
            "example": "2026-05-20"
          }
        },
        "required": ["title", "categoryId"]
      },
      "CreateTodoResponse": {
        "type": "object",
        "description": "할일 생성 성공 응답",
        "properties": {
          "todo": {
            "$ref": "#/components/schemas/Todo"
          }
        },
        "required": ["todo"]
      },
      "TodosResponse": {
        "type": "object",
        "description": "할일 목록 응답",
        "properties": {
          "todos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Todo"
            },
            "description": "필터 조건에 맞는 할일 목록 (생성일 내림차순)"
          }
        },
        "required": ["todos"]
      },
      "UpdateTodoRequest": {
        "type": "object",
        "description": "할일 수정 요청 본문. 수정할 필드만 포함하면 됩니다.",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "description": "변경할 할일 제목 (최대 255자)",
            "example": "주간 보고서 작성 (수정됨)"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "description": "변경할 카테고리 ID",
            "example": "550e8400-e29b-41d4-a716-446655440011"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "변경할 상세 설명 (null 전달 시 초기화)",
            "example": "화요일 오전까지로 변경"
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "description": "변경할 종료 예정일 (null 전달 시 초기화, YYYY-MM-DD)",
            "example": "2026-05-21"
          }
        }
      },
      "UpdateTodoResponse": {
        "type": "object",
        "description": "할일 수정 성공 응답",
        "properties": {
          "todo": {
            "$ref": "#/components/schemas/Todo"
          }
        },
        "required": ["todo"]
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "인증 토큰이 없거나 유효하지 않은 경우",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "UNAUTHORIZED": {
                "summary": "Authorization 헤더 누락",
                "value": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "message": "인증 토큰이 필요합니다. Authorization: Bearer <token> 헤더를 포함해주세요."
                  }
                }
              },
              "INVALID_TOKEN": {
                "summary": "토큰 검증 실패 또는 만료",
                "value": {
                  "error": {
                    "code": "INVALID_TOKEN",
                    "message": "토큰이 유효하지 않거나 만료되었습니다. 다시 로그인해주세요."
                  }
                }
              }
            }
          }
        }
      },
...
      "InvalidInputError": {
        "description": "요청 입력값이 유효하지 않은 경우",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "INVALID_INPUT",
                "message": "필수 입력 항목이 누락되었거나 형식이 올바르지 않습니다."
              }
            }
          }
        }
      },
      "InternalServerError": {
        "description": "서버 내부 오류",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "INTERNAL_SERVER_ERROR",
                "message": "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/todos": {
      "get": {
        "tags": ["Todos"],
        "summary": "할일 목록 조회",
        "description": "인증된 사용자의 할일 목록을 조회합니다. 쿼리 파라미터로 필터를 단독 또는 조합하여 적용할 수 있습니다.\n\n**필터 옵션 (BR-13):**\n- `categoryId`: 특정 카테고리의 할일만 조회\n- `startDate` / `endDate`: 종료 예정일(dueDate) 기간 필터\n- `isCompleted`: 완료 여부 필터 (true/false, 미지정 시 전체)\n\n**기한 초과 (BR-12):** `dueDate < 오늘 날짜 AND isCompleted = false`인 할일의 `overdue` 필드는 true로 반환됩니다.",
        "operationId": "getTodos",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "description": "특정 카테고리 ID로 필터링",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "550e8400-e29b-41d4-a716-446655440010"
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "description": "종료 예정일 범위 시작일 (YYYY-MM-DD). endDate와 함께 사용 권장",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "example": "2026-05-01"
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "description": "종료 예정일 범위 종료일 (YYYY-MM-DD). startDate와 함께 사용 권장",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "example": "2026-05-31"
          },
          {
            "name": "isCompleted",
            "in": "query",
            "required": false,
            "description": "완료 여부 필터. true: 완료된 할일만, false: 미완료 할일만, 미지정: 전체",
            "schema": {
              "type": "boolean"
            },
            "example": false
          }
        ],
        "responses": {
          "200": {
            "description": "할일 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodosResponse"
                },
                "examples": {
                  "전체_조회": {
                    "summary": "전체 할일 목록",
                    "value": {
                      "todos": [
                        {
                          "todoId": "550e8400-e29b-41d4-a716-446655440100",
                          "userId": "550e8400-e29b-41d4-a716-446655440001",
                          "categoryId": "550e8400-e29b-41d4-a716-446655440010",
                          "title": "주간 보고서 작성",
                          "description": "월요일 오전 9시까지 제출",
                          "dueDate": "2026-05-20",
                          "isCompleted": false,
                          "completedAt": null,
                          "overdue": false,
                          "createdAt": "2026-05-13T10:30:00.000Z",
                          "updatedAt": "2026-05-13T10:30:00.000Z"
                        },
                        {
                          "todoId": "550e8400-e29b-41d4-a716-446655440101",
                          "userId": "550e8400-e29b-41d4-a716-446655440001",
                          "categoryId": "550e8400-e29b-41d4-a716-446655440011",
                          "title": "장보기",
                          "description": null,
                          "dueDate": "2026-05-10",
                          "isCompleted": false,
                          "completedAt": null,
                          "overdue": true,
                          "createdAt": "2026-05-08T09:00:00.000Z",
                          "updatedAt": "2026-05-08T09:00:00.000Z"
                        }
                      ]
                    }
                  },
                  "완료된_할일_조회": {
                    "summary": "완료된 할일 목록",
                    "value": {
                      "todos": [
                        {
                          "todoId": "550e8400-e29b-41d4-a716-446655440102",
                          "userId": "550e8400-e29b-41d4-a716-446655440001",
                          "categoryId": "550e8400-e29b-41d4-a716-446655440010",
                          "title": "팀 미팅 준비",
                          "description": "발표 자료 준비",
                          "dueDate": "2026-05-12",
                          "isCompleted": true,
                          "completedAt": "2026-05-12T08:30:00.000Z",
                          "overdue": false,
                          "createdAt": "2026-05-10T11:00:00.000Z",
                          "updatedAt": "2026-05-12T08:30:00.000Z"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": ["Todos"],
        "summary": "할일 생성",
        "description": "인증된 사용자의 새로운 할일을 생성합니다.\n\n**비즈니스 규칙:**\n- 제목(title)과 카테고리(categoryId)는 필수입니다 (BR-09)\n- 설명(description)과 종료 예정일(dueDate)은 선택 사항입니다\n- 할일은 인증된 사용자에게 귀속됩니다 (BR-10)",
        "operationId": "createTodo",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTodoRequest"
              },
              "examples": {
                "전체_필드": {
                  "summary": "전체 필드 입력",
                  "value": {
                    "title": "주간 보고서 작성",
                    "categoryId": "550e8400-e29b-41d4-a716-446655440010",
                    "description": "월요일 오전 9시까지 팀장에게 제출",
                    "dueDate": "2026-05-20"
                  }
                },
                "필수_필드만": {
                  "summary": "필수 필드(title, categoryId)만 입력",
                  "value": {
                    "title": "장보기",
                    "categoryId": "550e8400-e29b-41d4-a716-446655440011"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "할일 생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTodoResponse"
                },
                "example": {
                  "todo": {
                    "todoId": "550e8400-e29b-41d4-a716-446655440100",
                    "userId": "550e8400-e29b-41d4-a716-446655440001",
                    "categoryId": "550e8400-e29b-41d4-a716-446655440010",
                    "title": "주간 보고서 작성",
                    "description": "월요일 오전 9시까지 팀장에게 제출",
                    "dueDate": "2026-05-20",
                    "isCompleted": false,
                    "completedAt": null,
                    "overdue": false,
                    "createdAt": "2026-05-13T10:30:00.000Z",
                    "updatedAt": "2026-05-13T10:30:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "필수 필드 누락 또는 유효성 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "title_누락": {
                    "summary": "제목 누락",
                    "value": {
                      "error": {
                        "code": "INVALID_INPUT",
                        "message": "할일 제목(title)은 필수 입력 항목입니다."
                      }
                    }
                  },
                  "categoryId_누락": {
                    "summary": "카테고리 ID 누락",
                    "value": {
                      "error": {
                        "code": "INVALID_INPUT",
                        "message": "카테고리(categoryId)는 필수 입력 항목입니다."
                      }
                    }
                  },
                  "dueDate_형식_오류": {
                    "summary": "날짜 형식 오류",
                    "value": {
                      "error": {
                        "code": "INVALID_INPUT",
                        "message": "종료 예정일(dueDate)은 YYYY-MM-DD 형식이어야 합니다."
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "지정한 카테고리 ID가 존재하지 않는 경우",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "NOT_FOUND",
                    "message": "지정한 카테고리를 찾을 수 없습니다."
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/todos/{todoId}": {
      "patch": {
        "tags": ["Todos"],
        "summary": "할일 수정",
        "description": "인증된 사용자가 본인이 소유한 할일을 수정합니다. 수정할 필드만 포함하면 됩니다.\n\n**비즈니스 규칙:**\n- 본인이 소유한 할일만 수정할 수 있습니다 (BR-10)\n- 제목, 설명, 종료 예정일, 카테고리를 수정할 수 있습니다",
        "operationId": "updateTodo",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "todoId",
            "in": "path",
            "required": true,
            "description": "수정할 할일의 UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "550e8400-e29b-41d4-a716-446655440100"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTodoRequest"
              },
              "examples": {
                "제목_수정": {
                  "summary": "제목 수정",
                  "value": {
                    "title": "주간 보고서 작성 (긴급)"
                  }
                },
                "전체_수정": {
                  "summary": "전체 필드 수정",
                  "value": {
                    "title": "주간 보고서 작성 (수정됨)",
                    "categoryId": "550e8400-e29b-41d4-a716-446655440011",
                    "description": "화요일 오전까지로 일정 변경",
                    "dueDate": "2026-05-21"
                  }
                },
                "설명_초기화": {
                  "summary": "설명 초기화 (null 전달)",
                  "value": {
                    "description": null
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "할일 수정 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateTodoResponse"
                },
                "example": {
                  "todo": {
                    "todoId": "550e8400-e29b-41d4-a716-446655440100",
                    "userId": "550e8400-e29b-41d4-a716-446655440001",
                    "categoryId": "550e8400-e29b-41d4-a716-446655440011",
                    "title": "주간 보고서 작성 (수정됨)",
                    "description": "화요일 오전까지로 일정 변경",
                    "dueDate": "2026-05-21",
                    "isCompleted": false,
                    "completedAt": null,
                    "overdue": false,
                    "createdAt": "2026-05-13T10:30:00.000Z",
                    "updatedAt": "2026-05-13T12:00:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidInputError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "description": "타인이 소유한 할일 수정 시도",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "FORBIDDEN",
                    "message": "해당 할일에 접근할 권한이 없습니다."
                  }
                }
              }
            }
          },
          "404": {
            "description": "할일 또는 지정한 카테고리를 찾을 수 없는 경우",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "할일_없음": {
                    "summary": "할일을 찾을 수 없음",
                    "value": {
                      "error": {
                        "code": "NOT_FOUND",
                        "message": "해당 할일을 찾을 수 없습니다."
                      }
                    }
                  },
                  "카테고리_없음": {
                    "summary": "지정한 카테고리를 찾을 수 없음",
                    "value": {
                      "error": {
                        "code": "NOT_FOUND",
                        "message": "지정한 카테고리를 찾을 수 없습니다."
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": ["Todos"],
        "summary": "할일 삭제",
        "description": "인증된 사용자가 본인이 소유한 할일을 영구적으로 삭제합니다.\n\n**비즈니스 규칙:**\n- 본인이 소유한 할일만 삭제할 수 있습니다 (BR-10)",
        "operationId": "deleteTodo",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "todoId",
            "in": "path",
            "required": true,
            "description": "삭제할 할일의 UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "550e8400-e29b-41d4-a716-446655440100"
          }
        ],
        "responses": {
          "200": {
            "description": "할일 삭제 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "삭제 성공 메시지",
                      "example": "할일이 삭제되었습니다."
                    }
                  },
                  "required": ["message"]
                },
                "example": {
                  "message": "할일이 삭제되었습니다."
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "description": "타인이 소유한 할일 삭제 시도",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "FORBIDDEN",
                    "message": "해당 할일에 접근할 권한이 없습니다."
                  }
                }
              }
            }
          },
          "404": {
            "description": "삭제할 할일을 찾을 수 없는 경우",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "NOT_FOUND",
                    "message": "해당 할일을 찾을 수 없습니다."
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  }
}